Se cambio funcion de lcd de capa
