+++ SistemaRiego/src/Aplicacion/SistemaRiego.c
@@ -100,17 +100,3 @@ void TimerEvent(){
 
 }
 
-void Display_lcd( char *msg , char renglon , char posicion )
-{
-	unsigned char i ;
-
-	if( renglon )
-        posicion = posicion + 0xc0 ;
-	else
-		posicion = posicion + 0x80 ;
-
-	Dato_LCD( posicion , LCD_CONTROL );
-	for( i = 0 ; msg[ i ] != '\0' ; i++ )
-		Dato_LCD( msg[ i ] , LCD_DATA );
-
-}
