+++ SistemaRiego/src/Primitivas/lcd.c
@@ -1,0 +1,24 @@
+/*
+ * lcd.c
+
+ *
+ *  Created on: 18 de nov. de 2016
+ *      Author: Martin
+ */
+#include "infotronic.h"
+#include "lcd.h"
+
+void Display_lcd( char *msg , char renglon , char posicion )
+{
+	unsigned char i ;
+
+	if( renglon )
+        posicion = posicion + 0xc0 ;
+	else
+		posicion = posicion + 0x80 ;
+
+	Dato_LCD( posicion , LCD_CONTROL );
+	for( i = 0 ; msg[ i ] != '\0' ; i++ )
+		Dato_LCD( msg[ i ] , LCD_DATA );
+
+}
